requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
cffi==1.16.0
construct==2.10.70
lxml==5.2.2
MouseInfo==0.1.3
PyAutoGUI==0.9.54
pycparser==2.22
pycryptodomex==3.20.0
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==0.1.30
python-dotenv==1.0.1
pytweening==1.2.0
